==> S1b_Off-by-one_of_major_codes_supplement.R <==
#S1b_Plotting_off-by-one_codes
#Pull top code from any sample and pull the off_by_one codes for that code 

#note -- this code is intended to be interacted with manually...  need to set the sample ID manually in line 24

# This code requires the samples_names.txt file in the working directory (also used in S1)
# and the .codes files generated by MultiCodes.pl (Wetmore et al) in a subfolder named "raw_data"

#~~~~~~~~~~
rm(list = ls()) #clear environment
library(tidyverse)
dir.create("Off_by_ones") 

#read sample names matrix with the base of the original file name in column 2 and new simple sample name in column 1
SampleNameTable <- read_table2("sample_names.txt", col_names = TRUE)

#make list of new names
SampleName <- SampleNameTable %>% pull(new_name)
numsamp <- length(SampleName)  #how many samples in set
sampleindex <- 1:numsamp   #make array of sampleindex numbers corresponding to number of samples

#don't loop, just set an x manually and play with one sample in 'sampleindex' at a time.  for (x in sampleindex) 
  
  x <- 30  #(manually set x to pull different LN quadrant samples)
    samp <- paste(SampleNameTable[x,2]) #original file name
    name <- paste(SampleNameTable[x,1]) #revised simple name
    
  
rawCodes <- read_table2(paste("raw_data/", samp, ".codes", sep=""), col_names = TRUE)
rawCodes  <- rawCodes %>% rename(counts = 2) %>% arrange(desc(counts))

bigCodes <- rawCodes %>% filter(counts > 99)

    i=1
    rmrows <- agrep(rawCodes$barcode[i], rawCodes$barcode, 1, value = FALSE)
    Off_by_one_1 <- rawCodes[rmrows,]
    bigcode <- as.numeric(Off_by_one_1[1,2])
    mismatches <- as.numeric(sum(Off_by_one_1[2:nrow(Off_by_one_1),2]))
    fracMismatch <- mismatches/bigcode
    Off_by_one_1 <- mutate(Off_by_one_1, "fracOfBig" = Off_by_one_1$counts/bigcode)

a <- ggplot(Off_by_one_1, aes(x=barcode, y=fracOfBig)) + geom_point() + scale_y_log10()
a    
    

write_tsv(Off_by_one_1, paste("Off_by_ones/", name, "_topcode.errors", sep=""), na = "0", append = FALSE, col_names = TRUE, quote_escape = "double")
