==> S2_Venn_diagrams.R <==
#S2_venn_diagrams_of_barcodes

#Read in filtered code lists. Replace counts data with "TRUE".  Full join relevant sets by barcode.
#Replace "N/A" with "FALSE".  Then use euler function in eulerr to calculate and plot overlap. 

# requires "abortus_pool.txt" which is the set of mapped codes
# requires filtered data in "filtered_data" folder generated by S1_initial_processing

library(tidyverse)
library(eulerr)
rm(list = ls()) #clear environment

######Overlap between Inoculum sets (Figure S5A)
#read in filtered inoc sets
InocA <- read_table2(file="filtered_data/Inoc_x_A.codesCln", col_names = TRUE)
InocB <- read_table2(file="filtered_data/Inoc_x_B.codesCln", col_names = TRUE)
InocC <- read_table2(file="filtered_data/Inoc_x_C.codesCln", col_names = TRUE)

#replace counts with TRUE
InocA <- InocA[,1] %>% mutate('A' = TRUE)
InocB <- InocB[,1] %>% mutate('B' = TRUE)
InocC <- InocC[,1] %>% mutate('C' = TRUE)

# joins of pairs of inoc samples, and all three
AB <- full_join(InocA, InocB, by = "barcode")
BC <- full_join(InocB, InocC, by = "barcode")
AC <- full_join(InocA, InocC, by = "barcode")
ABC <- full_join(AB, InocC, by= "barcode")
#convert NAs to False
ABC[is.na(ABC)] <- FALSE

plot(euler(ABC[, 2:4], shape = "ellipse"), quantities = TRUE)
#export plot as EPS

######Overlap between Inoc, mapped sets and cow (PT) (for the venn diagram in Figure S5B)

#read in barcode lists for Inoc (made with S1_initial processing) and mapped set
Inoc<- read_table2("R_output/Inoc_barcode_list.txt", col_names = TRUE)  # full inoc set filtered off-by-one
Mapped <- read_table2("abortus_pool.txt", col_names = TRUE) # mapped codes
Mapped <- Mapped[,2] %>% rename("barcode" = rcbarcode) # keep just barcode
PT_wI <- read_table2("R_output/All_PT_LN_samples_without_inoc_filter.codes", col_names = TRUE)
PT <-PT_wI[,1] # keep just barcode

#add a column to each to indicate TRUE
Inoc_T <- mutate(Inoc, 'inoc' = T)
Mapped_T <- mutate(Mapped, 'map' = T)
PT_T <- mutate(PT, 'pt' = T)

#join all into one table 
Inoc_PT <- full_join(Inoc_T, PT_T, by = "barcode")
Inoc_Mapped <- full_join(Inoc_T, Mapped_T, by = "barcode")
PT_Mapped <- full_join(PT_T, Mapped_T, by = "barcode")
ALL <- full_join(Inoc_Mapped, PT_T, by= "barcode")
#convert NAs to False
ALL[is.na(ALL)] <- FALSE

plot(euler(ALL[, 2:4], shape = "ellipse"), quantities = TRUE)
#export plot as EPS
